arc: networkd: Clean up child process on daemon exit.
The child process (ip-helper) does not exit as expected when the parent process
is terminated. Left in a useless state, it continually spams the logs with read
failures. This patch explicitly stops watching the control socket before
quitting which allows the process to terminate.

BUG=chromium:921550
BUG=chromium:934644
TEST=flashed nocturne, manually verified log spam is now gone when service goes down.
TEST=various tast and autotest presubmit tests

Change-Id: I058deacdfa32d2a35324e2c09331e3ae69a07d71
Reviewed-on: https://chromium-review.googlesource.com/1558835
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>