shill: Remove unused |lower_binder_| code
Connection::SetupExcludedRoutes uses a hack involving |lower_binder_|
when there are routes to be excluded but the Connection in question
does not use per device tables. Since the only case where
IPConfig::exclusion_list is not empty is for VPNs--which already use
per device tables--this code is currently unused. As we are
planning on moving all Connections over to per device tables, it
makes sense simply to remove this code.

This code is also the only user of |lower_binder_|, so this change
also removes that member and all associated methods. While Binders are
still available for use, Connections themselves no longer on any
Binder. Connections that need to depend on other Connection instances
should mediate this dependency through the use of Manager or some
other class with a view of all the available Connection instances
instead.

BUG=chromium:941597, chromium:942657
TEST=-All unit tests are passing.
     -l2tp vpn working as before (no CHECKs are triggered)

Change-Id: I5dd8b13167248685554b7734256e278e2f0a0d2a
Reviewed-on: https://chromium-review.googlesource.com/1562221
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>