biod: Add retries for device init operations
This CL adds an additional retry to the critical
fingerprint MCU probing operations, which take place during
biod startup. The error can be reproduced using the stress tester
in CL:1556384 (patchset 11).
The same stress tester was used to verify that this solution.

Reads and ioctls to the fingerprint MCU (/dev/cros_fp)
have been shown to occasionally fail with ETIMEDOUT.
If a read/cmd fails during the startup of biod, it may be led to
believe that the system does not have a fingerprint sensor/mcu.
This renders the fingerprint system inoperable until a reboot is issued.
Failures during the runtime of biod are less critical,
since the operation can be reattempted by the user.

This patchset does not cover the entire startup process.
A followup patchset will enable retries for the remaining
startup operations.

BUG=b:128011695
TEST=stress test to ensure 2 attempts is sufficient CL:1556384
TEST=implementation verified with stress tester in CL:1556384 PS 21
TEST=unit tests, emerged, deployed to nocturne
TEST=enrolled and used fingerprint

Change-Id: Ie0b849f9884f81fe75900ffd4b2652d1e1ef2e9d
Reviewed-on: https://chromium-review.googlesource.com/1559501
Commit-Ready: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>