vm_tools: concierge: Add workaround for crbug.com/909793
Add a workaround for crbug.com/909793 to pin the vcpu threads in crosvm
to the little cores.  We're still not sure what the root cause is but
this should at least allow people to use VMs without running into
lockups and undefined instruction errors.

This should be reverted once we find and fix the root cause.

BUG=chromium:909793
TEST=`dd if=/dev/urandom bs=4M count=1K of=foo.bin` succeeds without
     locking up the vm or causing undefined instruction errors

Change-Id: I2c5ce702891af0c60f3ba0c0afe4588f4f1c38b2
Reviewed-on: https://chromium-review.googlesource.com/1562464
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>