attestation: Add getters to handle all access of EK pubkey/cert
Currently we have serveral place in codebase that are doing the same
thing. Gather the EK public key or EK cert in protobuf database. If
it doesn't exists, it will get them manually from tpm utility and store
to the database (the old code did this but didn't call SaveDatabase()
after set the value).

In this CL, I add a utility that doing check the database, or get the EK
public key/cert from tpm_utility. But it doesn't store the value to
database due to partially set the database is messy.

BUG=chromium:910519
TEST=passed unit test

Change-Id: I40dfd8c2ff7a9e690112ec1a40a8657a63d5c252
Reviewed-on: https://chromium-review.googlesource.com/1530140
Commit-Ready: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>