shill: update setuid whitelist
Previously, this whitelist was in an insecure state, since a compromised
shill could switch to one of the unconstrained whitelisted UIDs (dhcp,
logger, nobody) and then switch to any UID on the system. Eliminate this
security hole by enforcing that none of shill's child UIDs are able to
switch to other users on the system.

BUG=chromium:845640
TEST=manual tests with SafeSetID LSM

Change-Id: I28afbcc2cd1109185e1f7772a15fcc0768a8ddeb
Reviewed-on: https://chromium-review.googlesource.com/1565213
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>