shill: vpn_service: Define GetUnderlyingConnection
VPNService currently depends on Connection::GetCarrierConnection to
determine what physical connection is used to send
traffic. GetCarrierConnection, however, does not actually work as
expected and instead simply returns the VPN Connection. Since
|lower_binder_| code is being removed and VPNService is the only
client of GetCarrierConnection, the functionality of determining what
connection a VPN will use is moved to VPNService.

In the future, it may be that a VPN may use different underlying
connections for different traffic. This method could be modified to
return a Connection given a set of traffic parameters or to return a
list of Connections and their associated traffic parameters, rather
than unconditionally returning a single Connection.

BUG=chromium:942657
TEST=-All unit tests are passing.
     -In chrome://network, the "PhysicalTechnology" property for a
      shill vpn interface now shows the actual physical technology.

Change-Id: I23f3e7e005fafb4c7e0bbbca22e16cbdbcb4ce89
Reviewed-on: https://chromium-review.googlesource.com/1564988
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>