arc: validate decoded output video by golden frame-wise MD5
VideoFrame is implemented for converting output frame data into
I420 pixel format cropping visible window. In order to compute MD5
and validate with golden one in frame-wise, and ouput to raw file
if needed.

For ARC++N, output buffer will return YUV_420_FLEXIBLE as color
format which may be either NV12 or YV12 underneath. In this case,
we need to try converting from both formats and match the golden
MD5 to find out the right format. This requires to be done in the
first output frame and then adopt to the whole stream.

Golden frame-wise MD5 is read from the text file which should be
named as <test-input-name>.frames.md5, and located in the same
directory.

BUG=b:112677333
TEST=run arcvideodecoder_test on Eve(P) and Coral(N)

Change-Id: I0b14399c0b08c7860bc93928657c511cef6cc3f3
Reviewed-on: https://chromium-review.googlesource.com/1565830
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>