crash: Deshell crash_sender completely
This removes the crash_sender.sh script and moves all the functionality
remaining in it into crash_sender_util.cc. The 2 main parts of it are
extracting all of the metadata and populating a FormData object with it;
and then also performing the actual HTTP request itself.

There is additional cleanup work that can be done; but I'm putting TODOs
in for a lot of that to keep this CL smaller. I'm also holding off on
moving functionality out of the autotest and into here directly to keep
the size smaller as well.

BUG=chromium:391887
TEST=Unit tests, autotest pass, upload works manually
Cq-Depend: chromium:1584028, chromium:1584418

Change-Id: I49fdc65ed85a3827d5b9478319883381f975e29e
Reviewed-on: https://chromium-review.googlesource.com/1584434
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>