biod: Switch to brillo's log file management fn
Brillo's version of UpdateLogSymlinks is nearly identical
to the local version. Switching to brillo's version adds
the errno string to the error message, but removes the sanity
check that all files are in the same directory.
In this case, the sanity check is not needed, since we explicitly
derive all paths from the same log directory path.
Brillo's GetTimeAsLogString is functionally equivalent to
the local GetTimeAsString.

The UpdateLogSymlinks code in biod was added in 2016 and
brillo's version was added in 2017.
Other projects have already made the switch to using brillo's version.

BUG=chromium:952548
TEST=deployed and manually verified

Change-Id: I5821861b2fecdba067e89db2a03ddbc569ef5031
Reviewed-on: https://chromium-review.googlesource.com/1565774
Commit-Ready: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>