shill: connection: Use per-device tables for all connections
This change causes Connection to use a per-device table for all
connections, and sets up routing rules such that the routing behavior
complies with Shill's priority rules. In particular, VPNs must be used
when applicable, and priorities of physical interfaces should follow
Shill's Service sorting order.

This implementation, as with the previous routing setup, does not
adequately handle the case where two distinct interfaces have the same
subnet. Future changes that allow for socket marking and routing based
on packet marks will be needed to resolve this issue.

Note that there is an existing issue (crbug.com/938629) where a
non-default physical interface is used for all v4 or v6 traffic when
the default physical interface does not have connectivity for that IP
version. This CL does not change the existing behavior with respect to
that issue.

BUG=chromium:941597
TEST=-All unit tests are passing.
     -Manually verified routing tables/rules.
     -Manually ensured that new Chrome connections use the expected
      (i.e. highest priority) interface.
     -Manually verified with `ip route get` and `ip -6 route get` that
      source address selection uses the highest priority interface for
      both v4 and v6 when we have multiple connected interfaces.
     -OpenVPN through Android works as expected.
     -Play store is still able to search and download apps
       (i.e. can access the network).
     -`sudo apt install emacs` using Crostini succeeds.
     -`test_that ${DUT} network_WiFi_Perf.ht40` succeeds.
     -`test_that ${DUT} network_VPNConnect.l2tpipsec_psk` succeeds.

Change-Id: I00664555d98004f3af86ac1ae71acbb49b721c58
Reviewed-on: https://chromium-review.googlesource.com/1590183
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>