shill: cellular: disable tests depending on CellularCapabilityClassic
This CL is part of a series to remove support of the legacy modem
manager (cromo) from shill.

As cromo is now deprecated, CellularCapabilityClassic and its derived
classes are no longer used. However, some of the unit tests depend on
CellularCapabilityClassic, which will need to be ported over to use
CellularCapabilityUniversal. This CL temporarily disables those tests so
that we can proceed with dropping CellularCapabilityClassic related
code. They will be re-enabled once they are ported over to use
CellularCapabilityUniversal instead.

BUG=chromium:954634
TEST=Run unit tests.

Change-Id: I0e256e0bda573de9b7df70d380be15ffb343b8d0
Reviewed-on: https://chromium-review.googlesource.com/1582193
Commit-Ready: Ben Chan <benchan@google.com>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>