biod: Change bio_crypto_init to dual log strategy
Change bio_crypto_init's logging behavior to
save historic logs and maintain LATEST/PREVIOUS
symbolic links to the two most recent.
This logging behavior is consistent with that of
biod and bio_fw_updater.

It should be noted that we are not changing the
logging directory to the standard /var/log/biod
yet. This is because bio_crypto_init is launched
from within a minijail, where a dedicated log
directory is bound for its sole use.
Changing to use the shared /var/log/biod directory
requires more evaluation of the isolation strategy.

BUG=chromium:954499
TEST=manually verified usage of new log scheme
Cq-Depend: chromium:1576710

Change-Id: I5a14dc56e5374f8fa94d93fcc910e5166b4395c9
Reviewed-on: https://chromium-review.googlesource.com/1574156
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>