common-mk: Allow tests to load .so from output dirs.
Some packages couldn't run test after unmerging the package. It was
becasue those referred a shared library that was installed in the system
directory before. After this change, tests can find .so files under the
build output directory hierarchy, which are built at the same build.
In GYP, we had these rpath settings so that built tests runs with the
shared library that was in its dependency.

BUG=chromium:954246
TEST=emerge-amd64-generic -C metrics; FEATURES=test emerge-amd64-generic metrics

Change-Id: Ibe217ba554f16948b262ed969f4795a019d28316
Reviewed-on: https://chromium-review.googlesource.com/1577231
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>