vm_tools: p9: Add unit tests
Add a bunch of tests for the 9p server code.  This isn't an exhaustive
list of tests but it covers most of the basic functionality needed to
consider the server to be functional.

Also clean up the handling of the lcreate message so that it is
consistent with the spec.

BUG=none
TEST=unit tests

Change-Id: I7ecd5718659727e80fe4c75d2e8c044b2af44d1d
Reviewed-on: https://chromium-review.googlesource.com/1594533
Commit-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>