diagnostics: fix CANCEL command hanging wilco_dtc_supportd
Add kill to the seccomp filter for wilco_dtc_supportd. Without
this, base::Process::Terminate() is blocked inside the minijail,
and attempting to execute a GetRoutineUpdate command with type
CANCEL will hang wilco_dtc_supportd. Also make sure the process
terminates when cancelled.

BUG=chromium:919151
TEST=execute diag --action=run_routine --routine=urandom, send
a cancel command to the running routine and make sure that a
response is returned.

Change-Id: I63f1a051d1c4272686ad2885ebd2db335c586f66
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1584964
Reviewed-by: Trent Begin <tbegin@chromium.org>