bluetooth: Avoid exposing advertisement interface before the chip is ready
There is a chance where the daemon is up waiting for the chip to
initiate, and this opens up a window for any client of newblued to register
for advertising. If a client call RegisterAdvertisement during this window,
newblued would configure the advertisement data but fail to enable
advertising due to the fact that the chip is not ready for use. This
leaves an adv set handle in libnewblue for no good. Besides, this
conflicts with the expectation of RegisterAdvertisement method, which is to
set the advertisement and enable it right away.
This also fixes a nit where std::move() is unnecessary with a raw
pointer.

BUG=None
TEST=Run unit tests

Change-Id: I96253f74dc4cfd587ace1cf1fdb75ee5d24f4cbc
Reviewed-on: https://chromium-review.googlesource.com/1586240
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>