shill: Modify metric selection process
The existing logic for determining the metrics of various Connection
instances has a few assumptions that limit the flexibility of the
solution. In particular, the code assumes that Connections only need a
single metric/priority for their routes/rules, which is already false.

This change modifies the metric selection logic in Manager to allow
for a Connection to create rules with multiple different priorities
(which is already done for VPN Connections using any form of
blacklisting). It also adds a parameter to the Connection::SetMetric
method which informs the Connection on whether or not it is the
primary physical device, as the assigned metric alone is insufficient
to determine this information. While this could be avoided by having an
explicit starting metric value for physical devices, this hardcodes a
maximum number of possible active VPN connections (and of high-priority
rules in general).

BUG=chromium:941597
TEST=-All unit tests are passing.
     -Manually verified with `ip rule && ip route` that priorities
      and metrics are being assigned as expected.

Change-Id: I4d260c84458cc3d35f1419da2f2922a67bb330f9
Reviewed-on: https://chromium-review.googlesource.com/1590182
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>