crash: Lock crash_sender lock file less
Don't hold the crash_sender lock file when not necessary. In particular,
release the lock file while sleeping.

This ensures that we don't block user initated uploads (which we want to
be fast, since the user asked for them) and login-crash-loop uploads
(which we need to be fast because the user may power-down the machine
soon) on deliberately slow background uploads.

This change is pretty straightforward. Unfortunately, the unit tests
ended up being quite ugly due to the semantics of file locking -- if the
current process holds the lock on the file, it will see the file as
unlocked. So I have to keep spawning other processes to lock the file
and/or test if the file is locked.

BUG=chromium:857317,chromium:951230
TEST=Ran manual crash upload while background upload was sleeping;
confirmed manual crash uploaded worked. Held lock, confirmed that
background upload wouldn't start. Grabbed lock during sleep, confirmed
that background upload stopped.
Cq-Depend: chromium:1595487

Change-Id: Ib21a69baf4bf2b3816142d075ce4861d22d11eb8
Reviewed-on: https://chromium-review.googlesource.com/1595465
Commit-Ready: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>