shill: cellular: remove obsolete 'Cellular.SupportedCarriers' property
This CL is part of a series to remove support of the legacy modem
manager (cromo) from shill.

The Cellular.SupportedCarriers property of a shill.Device has been
deprecated along with the deprecation of the cromo service. CL:1587007
removed the only usage of the property from Chrome.

BUG=chromium:954634
TEST=Run unit tests.

Change-Id: I13debbabe26e4e1a14bd3797b5f330e70daa24c9
Reviewed-on: https://chromium-review.googlesource.com/1589628
Commit-Ready: Ben Chan <benchan@google.com>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>