cryptohome: fix an autotest test for TPM 1.2 devices
Failing test: platform_InitLoginPerfServer.
Failure reason: RE pattern defined in the test for fetching the latency of
                preparing for enrollment doesn't match the log in the daemon.

This is a quick fix to stop the testing from failing. We should figure out a
more robust way to fetch the latency and update the code later. See b:131658301
for more details.

BUG=b:131658301
TEST=test_that <DUT's IP> platform_InitLoginPerfServer

Change-Id: I6471c7f29a2712bb9984b3341518cce73224117d
Reviewed-on: https://chromium-review.googlesource.com/1589299
Commit-Ready: Wei-Cheng Xiao <garryxiao@chromium.org>
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>