power: Add logic to identify first run after boot.
Currently we do not have a way to identify the first instantiation of
powerd after boot. So add a file to run directory to identify the same.
Absence of this file indicates that this is the first run. Presence of
this file means powerd has ran before after boot.

BUG=b:35505689
TEST=None

Change-Id: I92e3c93abf15cc64a2d5d58f3421c08a987a43f2
Reviewed-on: https://chromium-review.googlesource.com/1592351
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>