shill: rtnl: Add RTNLMessage::ToString
Currently, when an RTNL message results in an error, the RTNLHandler
logs an error that contains almost no information about the message
itself. As these messages can sometimes contain information that is
crucial to figuring out what went wrong, this is a detriment to
debugging.

BUG=chromium:958674
TEST=All unit tests are passing.

Change-Id: I4cadf97f83e88a7ea19cfd007b2d77c6f71beabb
Reviewed-on: https://chromium-review.googlesource.com/1594973
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>