vm_tools: concierge: stop ExecuteIo() in disk operation on error
Do not try to continue and copy more data when CopyEntry() fails. Check
the status of the operation and bail out.

Also check whether the output directory is still valid in MarkFailed().
We could have already "taken over it", in which case the path is empty
and we'll log error about inability to delete directory for no reason.

BUG=b:130304110
TEST=Import an image, verify dispatcher registers it.

Change-Id: If24452451a7183b318d026493641f68682fc1c1c
Reviewed-on: https://chromium-review.googlesource.com/1616153
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>