kerberos: Add org.chromium.Kerberos.service
The service file allows the service to be started on the first D-Bus
call to it. Thus, no call to Upstart is necessary anymore and the
permissions for Chrome to call Upstart can be removed.

Also does a minor renaming of file paths.

Cq-Depend: chromium:1593087

BUG=chromium:932075
TEST=None

Change-Id: I996025d06ccc52a6234ca4a9e026e88ca36abcc2
Reviewed-on: https://chromium-review.googlesource.com/1593235
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>