shill: rtnl: Overload SendMessage to transfer RTNLMessage ownership
RTNLHandler::SendMessage currently does not allow for RTNLMessage
ownership to be transfered from the client to RTNLHandler. This
prevents RTNLHandler from having context of a particular RTNLMessage
when the corresponding RTNL response is received (without copying the
RTNLMessage, that is). This change provides an overload for
SendMessage that does allow for this.

This implementation is careful to only allow a certain window of
sequences to be stored at any given time. If a particular request
happens not to get a response (for whatever reason), this would
essentially leak the memory of that RTNLMessage. LOG-ing when this
occurs will allow us to determine why this occurred.

When clients use this SendMessage overload, if the message results in
an RTNL error, RTNLHandler will provide a clear an informative error
log.

BUG=chromium:958674
TEST=-All unit tests are passing.

Change-Id: I41ded8d621e0b4930f922148bfee740388e072b7
Reviewed-on: https://chromium-review.googlesource.com/1594974
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>