biod: Fix fmap area out of bounds check
This check was supposed to validate that the area being read is
not only safe to read, but also considered valid with respect
to the fmap size boundary.

An fmap header lays out the outline for an actual firmware image
when loaded in flash. The base specifies the memory address where
the firmware image starts, the size specifies the overall addressable
firmware image size, and an area's offset/size specifies where that
section resides with respect to the base, but still within the overall
firmware image size.

We already check that fmap->size is less than or equal to the file size.
This change checks that the area is within fmap->size.

area->size/offset <= fmap->size <= file.length

BUG=b:117182903
TEST=unit test with fmap that reports an area outside fmap size
     (when fmap size is 0 works)

Change-Id: I71e1b0038fa16cc25079a9c9fc474d64dc3fc569
Reviewed-on: https://chromium-review.googlesource.com/1594137
Commit-Ready: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>