power: Refactor: Simplify dark resume logic.
dark resume logic when written was expected to run on 3.8 kernel.
3.8 kernel did not have support for CLOCK_REALTIME_ALARM and
CLOCK_BOOTTIME_ALARM. As a result dark resume logic when need to sleep
for certain time have to pound on /dev/rtc directly to schedule a wakeup.
As we are not planning to run dark resume on older kernels, we can simply
instantiate SimpleAlarmTimer in dark_resume.cc if needed. So let us not
pass the suspend duration to suspender anymore.

There is one more reason for this refactor. SimpleAlarmTimer in the
backend use timerfd_* API. timerfd_* allow the kernel to manage a list of
events and have it program the RTC so the alarm will trigger for the
earliest event in the list. This avoids the need for user space
applications to coordinate in order to share the hardware. For timerfd_*
to work properly, no user daemon should touch /dev/rtc directly. But any
suspend_duration passed to suspender will result in a write to /dev/rtc.
So let us not do that.

BUG=b:35505689
TEST=unittest passes

Change-Id: I900e5d99f182c257de451c3316096b8cdb88af2f
Reviewed-on: https://chromium-review.googlesource.com/1594810
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>