crash: Remove FORCE_OFFICIAL and add --dev flag
This removes the FORCE_OFFICIAL env var as we don't need that anymore
with the C++ unit testing. We added a '--dev' flag to crash_sender which
will skip the test & official image checks and upload the crashes
anyways, in this case they will go to the crash-staging server instead
of the normal crash server.

BUG=chromium:927933, chromium:391887
TEST=Unit and autotests pass
Cq-Depend: chromium:1597175

Change-Id: I4729ab194d6917c67afde6c6390ba7a77f583a25
Reviewed-on: https://chromium-review.googlesource.com/1597972
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Tim Zheng <timzheng@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>