kerberos: Serialize accounts to disk
Adds code to save the list of accounts to disk and load them back in.
Accounts are loaded on startup and saved on every change. Changes are
expected to be relatively rare, so this is fine.

Right now, the file is saved to /tmp, which is a tmpfs, so it's cleared
after the daemon shuts down. In a future CL, the storage location will
be changed to /run/daemon-store/kerberosd/<hash>, which makes this CL
finally useful.

Also removes krb5cc and krb5_conf paths out of account data as it can
be recomputed from the principal hash on demand.

BUG=chromium:932075
TEST=Checked that the file is actually saved by nsentering the mount
     namespace of the process.

Change-Id: Ib432a57282da77e68137d6c3ab92e39e166b9f9b
Reviewed-on: https://chromium-review.googlesource.com/1599401
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>