mems_setup: libiio wrapper
This is the first commit to establish the mems_setup project.

mems_setup is meant to replace the existing shell scripts that setup
sensors at ChromeOS boot with a more maintainable C++ codebase.

The tool is based upon libiio, but at the same time as part of
platform2/, it attempts to reuse existing code and tools when feasible
and has a strong emphasis on testing.

This commit sets up the project and contains C++ wrappers over the C
libiio API. These largely exist for the purpose of supporting mock
objects for testing, but they also help workaround a few issues in
libiio itself.

This is a smaller CL to aid in review, and reviewers should refer to
CL:1592822 for the broader context in which this change exists.

BUG=chromium:937916
TEST=builds

Change-Id: Ida366fd1e6ddf862b01445c8758f8788252c1e5f
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1601731
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ben Chan <benchan@google.com>