vm_tools: concierge: notify dispatcher when stopping Plugin VMs
We need to notify dispatcher when concierge decides (or rather gets
request) to stop VM(s). We distinguish 2 cases: stopping individual VM
versus stopping all VMs because the service is being shut down. In the
former case we try to shut down the VM, whereas in the latter case we
attempt to suspend the VM as this operation is distinct and is faster
with Plugin VMs (there is no difference in Crostini case).

BUG=b:130303625
TEST=Try stopping VMs

Change-Id: I811712bc7e2f9907f3ad38e491797507146175e6
Reviewed-on: https://chromium-review.googlesource.com/1628367
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>