sealed_storage: create a wrapper for Unseal
To allow using the sealed_storage package in the components that
can link ChromeOS libraries (libchrome, libbrillo), but don't have
an easy way to include their header files, create a wrapper around
SealedStorage class interface that doesn't require including those
headers to use and can be used with components compiled by a
different compiler: extern "C" and no non-trivial types in the API.

It is expected to be a temporary solution, so both the old interface
and the new wrapper function are provided by libsealed_storage.
Development of a proper solution is tracked in b:132202992.

BUG=b:129982188
TEST=Unseal the data sealed with SealedStorage class through the new
     wrapper:
     - echo "test" > /tmp/data
     - sealed_storage_tool --seal --policy_pcr=6 \
                           --data=/tmp/data \
                           --blob=/mnt/stateful_partition/blob
     - sealed_storage_tool --unseal --policy_pcr=6 \
                           --wrapper --data=/tmp/data \
                           --blob=/mnt/stateful_partition/blob
     (succeeds, /tmp/data contains "test")

Change-Id: I1b192f934cc810644e03fbe2f4f6f36a6de74bbc
Reviewed-on: https://chromium-review.googlesource.com/1603282
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Yufeng Wang <yufengwang@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Yufeng Wang <yufengwang@google.com>