crash: Don't upload crash reports if OS is over 6 months old
This stops crash_sender from uploading any crash reports if the
timestamp on the /etc/lsb-release file is more than 6 months in the
past. This can be overridden with the --dev flag which was added
recently. This also puts a 'os_millis' value in the crash metadata file
so that after an OS upgrade we will then delete the old crash files.

BUG=chromium:915381
TEST=Unit tests pass, manually verified

Change-Id: Ica3c99bd77c65b0cc76a98765bc889a564049e59
Reviewed-on: https://chromium-review.googlesource.com/1610803
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>