kerberos: Add is_managed flag
Adds a flag to indicate whether a Kerberos account is managed or not.
This flag is necessary to relate accounts set by the KerberosAccounts
policy to the accounts known to the Kerberos daemon. For instance, if
KerberosAccounts sets accounts (A, B, C), but now C is removed from the
policy, we need to know that we have to delete C from the daemon. C
could have been added by the user as well.

BUG=chromium:952239
TEST=None

Change-Id: I81e0e61ca8aae7b36e65172204070fa221273642
Reviewed-on: https://chromium-review.googlesource.com/1610065
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>