cryptohome: Use / as the rootdir in mount-encrypted
In the absence of the MOUNT_ENCRYPTED_ROOT environment variable, the
code was using "" and "/" inconsistently. A previous change resulted
in some code seeing "" instead of "/", leading to the TPM firmware
update binary validation to fail (see bug). This change consistently
switches things over to "/" to fix the bug and hopefully reduce the
potential for confusion in the future.

BUG=chromium:960465
TEST=State-preserving TPM firmware update works (see bug).

Change-Id: Id1344ac8c80f130e9257e66efa9e47464cd51f68
Reviewed-on: https://chromium-review.googlesource.com/1610062
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>