shill: rtnl: Remove SendMessage overload that doesn't gain ownership
RTNLHandler::SendMessage was migrating from an interface that took
RTNLMessage* to one that takes unique_ptr<RTNLMessage> (and thus
allows RTNLHandler to gain ownership of the RTNLMessage). The method
was overloaded while moving clients over to the newer interface.

This change removes the RTNLMessage* overload, moving the last uses of
the old SendMessage method in RTNLHandler over to the new interface.

Cq-Depend: chrome-internal:1297994
BUG=chromium:958674
TEST=All unit tests are passing.

Change-Id: I5b9672c651b0e1557450d715e10df226bc491e41
Reviewed-on: https://chromium-review.googlesource.com/1610943
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>