kerberos: Use a plain vector to store accounts
Changes the storage class from an unordered map to a vector. This has
the advantage that the order of accounts stays stable. It would be
confusing for users if a recently added account showed up in the middle
of the list instead of in the end. Note that there are no performance
concerns. Only a handful accounts are expected to be added per user.
If it becomes a problem, a separate lookup table could be used.

Also uses a proto internally for storing account data. This simplifies
serialization as the data does not have to be copied around. Note that
AccountManager stores accounts in a vector instead of AccountDataList
internally since it's easier to iterate and erase from.

BUG=chromium:963351
TEST=Unit tests.

Change-Id: I1675cb0d27874a8e2f06db487ce634d407107765
Reviewed-on: https://chromium-review.googlesource.com/1610068
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: A Olsen <olsen@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>