smbprovider: Allow shares to be mounted multiple times
If the browser is restarted, either intentionally or due to a crash,
smbprovider will keep the original mount and the browser will attempt to
mount any saved shares when it restarts. For this to work as a user
expects, smbprovider needs to be able to mount a share more than once.

BUG=chromium:961183
TEST=Unit tests
TEST=Add share and browse using the files app

Change-Id: Iab1349e1ccc05d16fdcbfddc3257d25969119ced
Reviewed-on: https://chromium-review.googlesource.com/1614820
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>