ml: Treat a model as supported if it has metadata
Previously ML Service checks for the validity of a ModelId by checking
if the enum value is out-of-range. We expect to remove support from
models over time (including the current Smart Dim model soon). Therefore
the range check is not correct. This CL changes the definition of
"supported" model to be "has a ModelMetadata" entry.

We also document that models that become unsupported should have their
names prefixed with UNSUPPORTED_. The UNKNOWN model previously used as a
default value and for range checking purposes becomes UNSUPPORTED_ in
this way.

TEST=ebuild clean test
BUG=chromium:937063

Change-Id: I529119d4b56767326d8bb0594a9719ec53db9df7
Reviewed-on: https://chromium-review.googlesource.com/1614831
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Michael Martis <martis@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>